create new video call effects class
